requirements.txt
absl-py==1.1.0
aiofiles==22.1.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.8.1
altair==4.2.2
aniso8601==9.0.1
anyio==3.6.1
apturl==0.5.2
argcomplete==3.2.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
audioread==3.0.0
autopage==0.5.1
autopep8==1.7.0
aws-wsgi==0.2.7
backcall==0.2.0
base58==2.1.1
bcrypt==3.2.0
beartype==0.10.4
beautifulsoup4==4.11.1
bitarray==2.6.0
black==22.8.0
bleach==5.0.1
blinker==1.4
boto3==1.26.71
botocore==1.29.71
Brlapi==0.8.3
cached-path==1.3.2
cachetools==4.2.4
catboost==1.1.1
certifi==2022.12.7
cffi==1.15.1
cfn-flip==1.3.0
channels==4.0.0
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
cliff==4.0.0
cloudpickle==2.1.0
cmaes==0.8.2
cmd2==2.4.2
colorama==0.4.4
colorlog==6.7.0
cryptography==37.0.4
cupshelpers==1.0
cycler==0.11.0
cytoolz==0.12.0
datasets==2.9.0
dbus-python==1.2.18
debugpy==1.6.2
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.1
djangorestframework==3.14.0
dnspython==2.3.0
docker==5.0.3
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docstring-parser==0.15
duplicity==0.8.21
durationpy==0.6
EasyProcess==1.1
entrypoint2==1.1
entrypoints==0.4
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
executing==0.8.3
fastapi==0.79.0
fasteners==0.14.1
fastjsonschema==2.15.3
ffmpeg==1.4
filelock==3.13.1
fire==0.4.0
Flask==2.2.2
Flask-Mail==0.9.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==2.5.1
flatbuffers==1.12
flet==0.1.41
fonttools==4.34.4
frozenlist==1.3.1
fsspec==2023.1.0
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.10.0
google-api-python-client==1.12.11
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==1.44.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
graphviz==0.20.1
greenlet==3.0.3
grpcio==1.47.0
gunicorn==20.1.0
gym==0.25.2
gym-notices==0.0.8
h11==0.13.0
h5py==3.7.0
hexbytes==0.3.0
hjson==3.1.0
htmlmin==0.1.12
httplib2==0.20.2
huggingface-hub==0.12.0
idna==3.4
ImageHash==4.2.1
imbalanced-learn==0.9.1
importlib-metadata==5.0.0
instaloader==4.10
ipfshttpclient==0.8.0a2
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.8
jsondiff==2.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.17.0
jupyter-server-mathjax==0.2.6
jupyterlab==3.2.9
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
jupyterlab-widgets==1.1.1
kaggle==1.5.12
kappa==0.6.0
keras==2.9.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
kfp==1.8.13
kfp-pipeline-spec==0.1.16
kfp-server-api==1.8.5
kiwisolver==1.4.3
kubernetes==18.20.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==14.0.1
librosa==0.9.2
lightgbm==3.3.3
llvmlite==0.39.1
lockfile==0.12.2
louis==3.20.0
lru-dict==1.1.8
macaroonbakery==1.3.1
Markdown==3.3.7
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mdurl==0.1.2
mediapipe==0.8.10.1
missingno==0.5.1
mistune==0.8.4
monotonic==1.6
more-itertools==8.10.0
mrx-link==0.8.2
mss==6.1.0
multiaddr==0.0.9
multidict==6.0.2
multimethod==1.8
multiprocess==0.70.14
mypy-extensions==0.4.3
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbdime==3.1.1
nbformat==5.4.0
nes-py==8.2.1
nest-asyncio==1.5.5
netaddr==0.8.0
networkx==2.8.5
nltk==3.8.1
notebook==6.4.12
notebook-shim==0.1.0
numba==0.56.4
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.0
olefile==0.46
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
optuna==3.0.1
outcome==1.2.0
packaging==23.0
pandas==1.5.1
pandas-profiling==3.2.0
pandocfilters==1.5.0
paramiko==2.9.3
parsimonious==0.8.1
parso==0.8.3
pathspec==0.10.1
pathtools==0.1.2
pbr==5.10.0
pexpect==4.8.0
phik==0.12.2
pickleshare==0.7.5
Pillow==9.4.0
placebo==0.9.0
platformdirs==4.1.0
plotly==5.11.0
polars==0.18.4
pooch==1.6.0
prettytable==3.4.1
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.3
psutil==5.9.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.15.0
pycups==2.0.1
pydantic==1.9.1
pydeck==0.8.0
pygame==2.1.2
pyglet==1.5.21
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.6.1
Pympler==1.0.1
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyperclip==1.8.2
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
pytesseract==0.3.10
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-multipart==0.0.5
python-slugify==6.1.2
pytorch-pretrained-bert==0.6.2
pytube==15.0.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
pyxdg==0.27
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
regex==2022.10.31
reportlab==3.6.8
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
requirements-parser==0.5.0
resampy==0.4.2
responses==0.18.0
rich==13.3.1
rlp==2.0.1
rsa==4.8
s3transfer==0.6.0
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.11.2
SecretStorage==3.3.1
selenium==4.6.0
semver==2.13.0
Send2Trash==1.8.0
sentencepiece==0.1.97
sentry-sdk==1.15.0
seqeval==1.2.2
setproctitle==1.3.2
shell-gpt==0.4.1
simpletransformers==0.63.9
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soundfile==0.11.0
soupsieve==2.3.2.post1
spacepy==0.4.1
SQLAlchemy==1.4.39
sqlparse==0.4.3
stack-data==0.3.0
starlette==0.19.1
stevedore==4.0.0
streamlit==1.18.1
strip-hints==0.1.10
systemd-python==234
tabulate==0.8.10
tangled-up-in-unicode==0.2.0
tenacity==8.1.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflowjs==3.18.0
termcolor==1.1.0
terminado==0.15.0
text-unidecode==1.3
texttable==1.6.4
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
transformers==4.26.1
trio==0.21.0
trio-websocket==0.9.2
troposphere==4.5.3
typer==0.7.0
types-setuptools==65.3.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.14
urllib3-secure-extra==0.1.0
usb-creator==0.3.7
validators==0.20.0
varint==1.0.2
virtualenv==20.25.0
visions==0.7.4
wadllib==1.3.6
wandb==0.13.10
watchdog==2.2.1
wcwidth==0.2.5
web3==5.31.1
webencodings==0.5.1
websocket-client==1.3.3
websockets==9.1
Werkzeug==2.2.2
widgetsnbextension==3.6.1
wrapt==1.14.1
wsproto==1.1.0
xdg==5
xkit==0.0.0
xxhash==3.2.0
yarl==1.8.1
zappa==0.58.0
zipp==3.10.0
